Remove import cycle for query history dto mappers
