+++ extensions/ql-vscode/src/query-history/store/query-history-dto-mapper.ts
@@ -1,10 +1,7 @@
 import type { QueryHistoryInfo } from "../query-history-info";
 import type { QueryHistoryItemDto } from "./query-history-dto";
-import { QueryLanguageDto } from "./query-history-dto";
 import { mapQueryHistoryVariantAnalysisToDomainModel } from "./query-history-variant-analysis-dto-mapper";
 import { mapLocalQueryItemToDomainModel } from "./query-history-local-query-dto-mapper";
-import { QueryLanguage } from "../../common/query-language";
-import { assertNever } from "../../common/helpers-pure";
 
 export function mapQueryHistoryToDomainModel(
   queries: QueryHistoryItemDto[],
@@ -22,29 +19,4 @@ export function mapQueryHistoryToDomainModel(
       )}`,
     );
   });
-}
-
-export function mapQueryLanguageToDomainModel(
-  language: QueryLanguageDto,
-): QueryLanguage {
-  switch (language) {
-    case QueryLanguageDto.CSharp:
-      return QueryLanguage.CSharp;
-    case QueryLanguageDto.Cpp:
-      return QueryLanguage.Cpp;
-    case QueryLanguageDto.Go:
-      return QueryLanguage.Go;
-    case QueryLanguageDto.Java:
-      return QueryLanguage.Java;
-    case QueryLanguageDto.Javascript:
-      return QueryLanguage.Javascript;
-    case QueryLanguageDto.Python:
-      return QueryLanguage.Python;
-    case QueryLanguageDto.Ruby:
-      return QueryLanguage.Ruby;
-    case QueryLanguageDto.Swift:
-      return QueryLanguage.Swift;
-    default:
-      assertNever(language);
-  }
 }
